Moved pydantic schemas to separate module, added flask app, separate api and mounted to app
