+++ valometa/api/routes.py
@@ -4,31 +4,31 @@ from datetime import date
 from typing import List
 
 from fastapi import FastAPI
+from fastapi.middleware.wsgi import WSGIMiddleware
+from flask import Flask
 from pydantic import BaseModel
 from sqlalchemy import create_engine
 
+from valometa.api.schemas import DateRange, NumberMatchesDay
 from valometa.data import sqlite_db_path
 from valometa.utils.data import get_matches_per_day
 
 
-class DateRange(BaseModel):
-    date_begin: date
-    date_end: date
+api_app = FastAPI()
+flask_app = Flask(__name__)
+
+app = FastAPI(title="App Root")
+
+app.mount("/api", api_app)
+app.mount("/", WSGIMiddleware(flask_app))
 
 
-class NumberMatchesDay(BaseModel):
-    date_of_count: date
-    count: int
+@flask_app.get("/")
+def hello():
+    return 'Hello, World!'
 
 
-app = FastAPI()
-
-@app.get("/")
-def root():
-    return {'message': 'hello world'}
-
-
-@app.post("/matches-per-day")
+@api_app.post("/matches-per-day", response_model=List[NumberMatchesDay])
 def matches_per_day(date_range: DateRange) -> List[NumberMatchesDay]:
     engine = create_engine(f"sqlite:///{sqlite_db_path}")
     matches_df = (
